requirements.txt
Flask==2.0.2
google-python-cloud-debugger
google-cloud-logging